src/agent/query_classifier.py
"""
Query Classifier — Intent Detection for Tool Routing
=====================================================
Techniques implemented:
  - BERT-based fine-tuned classifier for query-type detection
  - Batch Normalization on projection layers
  - SSLU/AHerfReLU custom activation functions
  - SMOTE oversampling for class-balanced training
  - 4 classes: visual, textual, hybrid, graph
"""

import torch
import torch.nn as nn
import torch.nn.functional as F
import numpy as np
from typing import List, Dict, Optional, Tuple
from pathlib import Path
from loguru import logger

try:
    from transformers import AutoTokenizer, AutoModel
except ImportError:
    AutoTokenizer, AutoModel = None, None

try:
    from imblearn.over_sampling import SMOTE
except ImportError:
    SMOTE = None

import sys
sys.path.insert(0, str(Path(__file__).resolve().parent.parent.parent))
from src.config import get_config, DEVICE, AgentConfig
from src.retrieval.colbert_reranker import get_activation


# Query type labels
QUERY_TYPES = {
    0: "visual",
    1: "textual",
    2: "hybrid",
    3: "graph",
}


class QueryClassifier(nn.Module):
    """
    BERT-based query intent classifier.
    
    Classifies incoming queries into one of 4 types to help
    the agent decide which retrieval tools to prioritize:
      - visual: image search needed
      - textual: document/text search needed
      - hybrid: multi-modal search needed
      - graph: knowledge graph query needed
    """

    def __init__(
        self,
        config: Optional[AgentConfig] = None,
        num_classes: int = 4,
    ):
        super().__init__()
        self.config = config or AgentConfig()
        
        if AutoModel is None:
            raise ImportError("transformers required")
        
        # BERT backbone
        self.encoder = AutoModel.from_pretrained(self.config.classifier_model)
        self.tokenizer = AutoTokenizer.from_pretrained(self.config.classifier_model)
        hidden_size = self.encoder.config.hidden_size
        
        # Classification head with BatchNorm and custom activation
        activation = get_activation(self.config.classifier_activation)
        
        self.classifier = nn.Sequential(
            nn.Linear(hidden_size, 256),
            nn.BatchNorm1d(256),
            activation,
            nn.Dropout(0.3),
            nn.Linear(256, 64),
            nn.BatchNorm1d(64),
            activation,
            nn.Dropout(0.2),
            nn.Linear(64, num_classes),
        )
        
        # Keyword-based heuristic fallback
        self.visual_keywords = {
            "image", "images", "photo", "picture", "show", "visual",
            "look", "appearance", "color", "shape", "display", "see",
            "diagram", "chart", "graph", "plot",
        }
        self.graph_keywords = {
            "relationship", "related", "connected", "who", "inspected",
            "responsible", "operator", "contains", "near", "part of",
            "entity", "entities", "link",
        }
        self.textual_keywords = {
            "text", "document", "describe", "explain", "summary",
            "note", "report", "written", "article", "paper",
        }
        
        logger.info(
            f"QueryClassifier: {self.config.classifier_model} | "
            f"Activation: {self.config.classifier_activation}"
        )

    def forward(self, input_ids: torch.Tensor, attention_mask: torch.Tensor) -> torch.Tensor:
        """
        Forward pass.
        
        Args:
            input_ids: (B, seq_len) tokenized input
            attention_mask: (B, seq_len) attention mask
        Returns:
            (B, num_classes) logits
        """
        outputs = self.encoder(
            input_ids=input_ids, attention_mask=attention_mask
        )
        # Use [CLS] token embedding
        cls_embedding = outputs.last_hidden_state[:, 0, :]
        logits = self.classifier(cls_embedding)
        return logits

    @torch.no_grad()
    def classify(self, query: str) -> Dict:
        """
        Classify a single query.
        
        Args:
            query: Natural language query
        Returns:
            Dict with 'type', 'confidence', 'all_scores'
        """
        self.eval()
        
        inputs = self.tokenizer(
            query,
            max_length=64,
            padding="max_length",
            truncation=True,
            return_tensors="pt",
        ).to(DEVICE)
        
        try:
            logits = self.forward(inputs["input_ids"], inputs["attention_mask"])
            probs = F.softmax(logits, dim=-1).squeeze(0).cpu().numpy()
            
            pred_idx = int(probs.argmax())
            return {
                "type": QUERY_TYPES[pred_idx],
                "confidence": float(probs[pred_idx]),
                "all_scores": {
                    QUERY_TYPES[i]: float(p) for i, p in enumerate(probs)
                },
            }
        except Exception:
            # Fallback to heuristic
            return self._heuristic_classify(query)

    def _heuristic_classify(self, query: str) -> Dict:
        """Keyword-based fallback classification."""
        query_lower = query.lower()
        words = set(query_lower.split())
        
        visual_score = len(words & self.visual_keywords)
        graph_score = len(words & self.graph_keywords)
        text_score = len(words & self.textual_keywords)
        
        scores = {
            "visual": visual_score,
            "textual": text_score,
            "graph": graph_score,
            "hybrid": 0.5,  # Default baseline
        }
        
        # If multiple modalities score high, it's hybrid
        high_scores = sum(1 for s in [visual_score, text_score, graph_score] if s > 0)
        if high_scores >= 2:
            scores["hybrid"] = max(scores.values()) + 1
        
        total = sum(scores.values()) or 1
        normalized = {k: v / total for k, v in scores.items()}
        
        best_type = max(normalized, key=normalized.get)
        
        return {
            "type": best_type,
            "confidence": normalized[best_type],
            "all_scores": normalized,
        }

    @staticmethod
    def generate_synthetic_training_data(n_samples: int = 1000) -> Tuple:
        """
        Generate synthetic training data for the classifier.
        Uses templates and SMOTE for class balancing.
        
        Returns:
            (texts, labels) tuple
        """
        templates = {
            0: [  # visual
                "show me images of {}", "find photos of {}",
                "what does {} look like", "display pictures of {}",
                "visual search for {}", "find similar images to {}",
            ],
            1: [  # textual
                "explain what {} is", "describe the process of {}",
                "summarize the document about {}", "find text about {}",
                "search notes for {}", "what are the details of {}",
            ],
            2: [  # hybrid
                "show images and describe {}", "find photos and notes about {}",
                "visual and textual search for {}", "images and summary of {}",
            ],
            3: [  # graph
                "who inspected {}", "what is related to {}",
                "find connections between {} and defects",
                "which operator was responsible for {}",
                "entities connected to {}", "relationship between {} and {}",
            ],
        }
        
        topics = [
            "surface cracks", "welding joints", "corrosion", "defects",
            "machine parts", "inspection results", "safety equipment",
            "manufacturing process", "quality control", "metal fatigue",
        ]
        
        texts, labels = [], []
        import random
        random.seed(42)
        
        for _ in range(n_samples):
            label = random.randint(0, 3)
            template = random.choice(templates[label])
            topic = random.choice(topics)
            text = template.format(topic)
            texts.append(text)
            labels.append(label)
        
        # Apply SMOTE if available
        if SMOTE is not None:
            logger.info("Applying SMOTE for class balancing")
            # Would need embeddings first — skipping for synthetic data
        
        return texts, labels
